make SchemaDiffMigration to use versioning which is the same logic as OrmaMigration
